EuiModal[X] TS
