+++ src/components/modal/modal_body.test.tsx
@@ -2,9 +2,7 @@ import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test/required_props';
 
-import {
-  EuiModalBody,
-} from './modal_body';
+import { EuiModalBody } from './modal_body';
 
 test('renders EuiModalBody', () => {
   const component = <EuiModalBody {...requiredProps}>children</EuiModalBody>;
