+++ src/components/modal/modal_footer.test.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test/required_props';
 
-import {
-  EuiModalFooter,
-} from './modal_footer';
+import { EuiModalFooter } from './modal_footer';
 
 test('renders EuiModalFooter', () => {
-  const component = <EuiModalFooter {...requiredProps}>children</EuiModalFooter>;
+  const component = (
+    <EuiModalFooter {...requiredProps}>children</EuiModalFooter>
+  );
   expect(render(component)).toMatchSnapshot();
 });
