+++ src/components/modal/modal_footer.tsx
@@ -1,0 +1,14 @@
+import React, { FunctionComponent, HTMLAttributes } from 'react';
+import classnames from 'classnames';
+import { CommonProps } from '../common';
+
+export const EuiModalFooter: FunctionComponent<
+  HTMLAttributes<HTMLDivElement> & CommonProps
+> = ({ className, children, ...rest }) => {
+  const classes = classnames('euiModalFooter', className);
+  return (
+    <div className={classes} {...rest}>
+      {children}
+    </div>
+  );
+};
