--- src/components/modal/modal_header_title.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import classnames from 'classnames';
-import PropTypes from 'prop-types';
-
-export function EuiModalHeaderTitle({ className, children, ...rest }) {
-  const classes = classnames('euiModalHeader__title', className);
-  return (
-    <div className={classes} {...rest}>
-      { children }
-    </div>
-  );
-}
-
-EuiModalHeaderTitle.propTypes = {
-  className: PropTypes.string,
-  children: PropTypes.node
-};
