+++ src/components/modal/modal_header_title.test.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test/required_props';
 
-import {
-  EuiModalHeaderTitle,
-} from './modal_header_title';
+import { EuiModalHeaderTitle } from './modal_header_title';
 
 test('renders EuiModalHeaderTitle', () => {
-  const component = <EuiModalHeaderTitle {...requiredProps}>children</EuiModalHeaderTitle>;
+  const component = (
+    <EuiModalHeaderTitle {...requiredProps}>children</EuiModalHeaderTitle>
+  );
   expect(render(component)).toMatchSnapshot();
 });
